# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorResponse {
  error: string;
}

export interface MatterExtractionDesignPage {
  /** Data URL (base64) of the card design page photo */
  image: string;
  /** Customer note describing which matter goes on this page */
  command: string;
}

export interface MatterExtractionInput {
  /** Data URLs (base64) of the filled physical form photos */
  formPhotos: string[];
  /** Card design pages, each with its placement command */
  designPages: MatterExtractionDesignPage[];
}

export interface ExtractedProgramme {
  name: string;
  /** Date as written on the form (e.g. "Sunday, 14 Dec 2025") */
  date: string;
  /** Time as written (e.g. "7:30 PM") */
  time: string;
  venue: string;
  address: string;
}

export interface ExtractedPerson {
  name: string;
  fatherName: string;
  motherName: string;
  grandfatherName: string;
  grandmotherName: string;
}

export interface ExtractedMatter {
  brideName: string;
  groomName: string;
  bride: ExtractedPerson;
  groom: ExtractedPerson;
  /** Hosting family name/title line */
  familyTitle: string;
  nativePlace: string;
  residenceAddress: string;
  deities: string[];
  /** Relation word such as "weds", "विवाह", etc. */
  relationWord: string;
  /** Main invitation body / blessing line if present */
  invitationText: string;
  programmes: ExtractedProgramme[];
  /** Summary of placement instructions across design pages */
  designNotes: string;
  /** Detected primary language/script (English, Hindi, Marathi, Gujarati) */
  language: string;
}