# lib/api-zod/src/generated/types/extractedMatter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedPerson } from './extractedPerson';
import type { ExtractedProgramme } from './extractedProgramme';

export interface ExtractedMatter {
  brideName: string;
  groomName: string;
  bride: ExtractedPerson;
  groom: ExtractedPerson;
  /** Hosting family name/title line */
  familyTitle: string;
  nativePlace: string;
  residenceAddress: string;
  deities: string[];
  /** Relation word such as "weds", "विवाह", etc. */
  relationWord: string;
  /** Main invitation body / blessing line if present */
  invitationText: string;
  programmes: ExtractedProgramme[];
  /** Summary of placement instructions across design pages */
  designNotes: string;
  /** Detected primary language/script (English, Hindi, Marathi, Gujarati) */
  language: string;
}