# lib/api-zod/src/generated/types/extractedPerson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedPerson {
  name: string;
  fatherName: string;
  motherName: string;
  grandfatherName: string;
  grandmotherName: string;
}