# lib/api-zod/src/generated/types/extractedProgramme.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedProgramme {
  name: string;
  /** Date as written on the form (e.g. "Sunday, 14 Dec 2025") */
  date: string;
  /** Time as written (e.g. "7:30 PM") */
  time: string;
  venue: string;
  address: string;
}