# lib/api-zod/src/generated/types/matterExtractionDesignPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatterExtractionDesignPage {
  /** Data URL (base64) of the card design page photo */
  image: string;
  /** Customer note describing which matter goes on this page */
  command: string;
}