# lib/api-zod/src/generated/types/matterExtractionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatterExtractionDesignPage } from './matterExtractionDesignPage';

export interface MatterExtractionInput {
  /** Data URLs (base64) of the filled physical form photos */
  formPhotos: string[];
  /** Card design pages, each with its placement command */
  designPages: MatterExtractionDesignPage[];
}